moving common components used in design tokens to a shared folder
