+++ src/components/DesignTokens/Color/Color.stories.js
@@ -1,15 +1,6 @@
 import React from 'react';
-import styled from 'styled-components';
 import Color from './Color';
-import { color, fontSize, spacing } from '../../../styles';
-
-const Import = styled.code`
-  background: ${color.whiteSmoke};
-  display: block;
-  font-size: ${fontSize.sm};
-  text-align: center;
-  padding: ${spacing.sm};
-`;
+import { ImportCode } from '../utils';
 
 export default {
   title: 'Design Tokens|Colors',
@@ -19,7 +10,7 @@ export default {
 export const All = () => {
   return (
     <>
-      <Import>{`import { color } from '../../styles'`};</Import>
+      <ImportCode>{`import { color } from '../../styles'`};</ImportCode>
       <Color tokenGroup="globalColor" heading="Global" description="Lorem ipsum dolor sit amet" />
       <Color tokenGroup="atkColor" heading="ATK" description="Lorem ipsum dolor sit amet" />
       <Color tokenGroup="ccoColor" heading="CCO" description="Lorem ipsum dolor sit amet" />
