--- src/components/DesignTokens/TokenTable/index.js
@@ -1,1 +1,0 @@
-export { default } from './TokenTable';
