+++ src/components/DesignTokens/components/TokenTable.js
