+++ src/components/DesignTokens/components/index.js
@@ -1,0 +1,7 @@
+import TokenTable from './TokenTable';
+import ImportCode from './ImportCode';
+
+export {
+  TokenTable,
+  ImportCode,
+};
